+++ open_instruct/distillkit/__init__.py
@@ -12,14 +12,20 @@ from open_instruct.distillkit.compression import (
     torch_dtype_to_name,
     unpack_from_bytes,
 )
+from open_instruct.distillkit.losses import ForwardKLTopKOutput, forward_kl_topk_from_logprobs
+from open_instruct.distillkit.vllm_logprobs import extract_response_topk_from_prompt_logprobs, process_prompt_logprobs
 
 __all__ = [
     "DistributionQuantizationConfig",
+    "ForwardKLTopKOutput",
     "LogprobCompressor",
     "QuantizationBin",
     "SpecialTerm",
+    "extract_response_topk_from_prompt_logprobs",
+    "forward_kl_topk_from_logprobs",
     "pack_to_bytes",
     "parse_torch_dtype",
+    "process_prompt_logprobs",
     "torch_dtype_to_name",
     "torch_dtype_bit_width",
     "unpack_from_bytes",